Move functionality to RuleBoardWrapper for clarity
