+++ src/modules/RuleBoardWrapper/index.ts
@@ -1,0 +1,3 @@
+import RuleBoardWrapper from "./RuleBoardWrapper";
+
+export default RuleBoardWrapper;
